+++ drift/lib/src/query_builder.dart
@@ -12,6 +12,7 @@ export 'query_builder/expressions/variables.dart';
 
 export 'query_builder/schema/column_constraints.dart';
 export 'query_builder/schema/column.dart';
+export 'query_builder/schema/drop.dart';
 export 'query_builder/schema/entities.dart';
 export 'query_builder/schema/result_set.dart';
 export 'query_builder/schema/table.dart';