+++ drift/lib/src/query_builder/schema/entities.dart
@@ -1,10 +1,25 @@
 import '../compiler.dart';
+import '../statements/statement.dart';
 
 /// Some abstract schema entity that can be stored in a database. This includes
 /// tables, triggers, views, indexes, etc.
 abstract interface class DatabaseSchemaEntity {
   /// The (unalised) name of this entity in the database.
   String get entityName;
+}
+
+/// A statement that creates a [DatabaseSchemaEntity] (such as tables, views,
+/// triggers or indices).
+abstract base class CreateStatement<T extends DatabaseSchemaEntity>
+    extends SqlStatement {
+  /// The table, view, trigger or index to create.
+  final T entity;
+
+  /// Whether the [entity] should only be created if it doesn't exist alredy.
+  final bool ifNotExists;
+
+  /// @nodoc
+  CreateStatement(this.entity, {this.ifNotExists = false});
 }
 
 /// A sqlite trigger that's executed before, after or instead of a subset of
@@ -16,7 +31,7 @@ abstract interface class DatabaseSchemaEntity {
 ///
 /// [sqlite-docs]: https://sqlite.org/lang_createtrigger.html
 /// [sql-tut]: https://www.sqlitetutorial.net/sqlite-trigger/
-class Trigger extends DatabaseSchemaEntity {
+final class Trigger extends DatabaseSchemaEntity {
   @override
   final String entityName;
 
@@ -30,6 +45,66 @@ class Trigger extends DatabaseSchemaEntity {
 
   /// Creates a trigger backed by an [sql] string that's not dialect-specific.
   Trigger.simpleSql(this.entityName, String sql)
+      : generateDefinition =
+            ((compiler) => compiler.statement.buffer.write(sql));
+}
+
+/// Represents a `CREATE TRIGGER` statement in SQL.
+final class CreateTriggerStatement extends CreateStatement<Trigger> {
+  /// Create a statement that will `CREATE` the [entity] when issued.
+  CreateTriggerStatement(super.entity, {super.ifNotExists});
+
+  @override
+  void compileWith(StatementCompiler compiler) {
+    return compiler.addCreateTriggerStatement(this);
+  }
+}
+
+/// An index on a table.
+final class Index extends DatabaseSchemaEntity {
+  @override
+  final String entityName;
+
+  /// A function responsible for writing the `CREATE INDEX` definition given
+  /// a [StatementCompiler].
+  final void Function(StatementCompiler) generateDefinition;
+
+  /// Creates a trigger from its name and the (possibly dialect-specific)
+  /// definition generator.
+  Index(this.entityName, this.generateDefinition);
+
+  /// Creates an index backed by an [sql] string that's not dialect-specific.
+  Index.simpleSql(this.entityName, String sql)
+      : generateDefinition =
+            ((compiler) => compiler.statement.buffer.write(sql));
+}
+
+/// Represents a `CREATE INDEX` statement in SQL.
+final class CreateIndexStatement extends CreateStatement<Index> {
+  /// Create a statement that will `CREATE` the [entity] when issued.
+  CreateIndexStatement(super.entity, {super.ifNotExists});
+
+  @override
+  void compileWith(StatementCompiler compiler) {
+    return compiler.addCreateIndexStatement(this);
+  }
+}
+
+/// An index on a table.
+final class OnCreateQuery extends DatabaseSchemaEntity {
+  @override
+  final String entityName;
+
+  /// A function responsible for writing the query to run when creating the
+  /// database.
+  final void Function(StatementCompiler) generateDefinition;
+
+  /// Creates a trigger from its name and the (possibly dialect-specific)
+  /// definition generator.
+  OnCreateQuery(this.entityName, this.generateDefinition);
+
+  /// Creates an index backed by an [sql] string that's not dialect-specific.
+  OnCreateQuery.simpleSql(this.entityName, String sql)
       : generateDefinition =
             ((compiler) => compiler.statement.buffer.write(sql));
 }